chore: rename TypeRegistry to DtrTest, as it depends on dtr-test schemas and structure
